if文追加

+++ init.sh
@@ -9,6 +9,9 @@ cd ${DOTPATH}
 if [ ! -d ${DOTPATH}/downloads/vim-hybrid ]; then
   git clone git@github.com:w0ng/vim-hybrid.git ${DOTPATH}/downloads/vim-hybrid
 fi
-cp ${DOTPATH}/downloads/vim-hybrid/colors/hybrid.vim ${DOTPATH}/.vim/colors/
+
+if [ ! -f ${DOTPATH}/.vim/colors/hybrid.vim ]; then
+  cp ${DOTPATH}/downloads/vim-hybrid/colors/hybrid.vim ${DOTPATH}/.vim/colors/
+fi
 
 exit 0
